+++ bin/next-export
@@ -3,7 +3,7 @@ import { resolve, join } from 'path'
 import { existsSync } from 'fs'
 import { cpus } from 'os'
 import parseArgs from 'minimist'
-import exportApp from '../server/export'
+import exportApp from '../export'
 import { printAndExit } from '../lib/utils'
 
 const argv = parseArgs(process.argv.slice(2), {
