+++ package.json
@@ -38,7 +38,7 @@
     "test": "cross-env npm run testall || npm run testall",
     "coveralls": "nyc --instrument=false --source-map=false report --temp-directory=./coverage --reporter=text-lcov | coveralls",
     "flow": "flow check",
-    "lint": "standard 'bin/*' 'client/**/*.js' 'examples/**/*.js' 'lib/**/*.js' 'pages/**/*.js' 'server/**/*.js' 'build/**/*.js' 'test/**/*.js'",
+    "lint": "standard 'bin/*' 'client/**/*.js' 'examples/**/*.js' 'lib/**/*.js' 'pages/**/*.js' 'server/**/*.js' 'export/**/*.js' 'build/**/*.js' 'test/**/*.js'",
     "prepublish": "npm run release",
     "precommit": "lint-staged"
   },
@@ -65,6 +65,7 @@
     "@babel/runtime": "7.0.0-beta.42",
     "@babel/template": "7.0.0-beta.42",
     "ansi-html": "0.0.7",
+    "async-sema": "2.1.1",
     "babel-core": "7.0.0-bridge.0",
     "babel-loader": "8.0.0-beta.3",
     "babel-plugin-react-require": "3.0.0",
@@ -139,7 +140,6 @@
     "node-notifier": "5.1.2",
     "nyc": "11.2.1",
     "progress": "2.0.0",
-    "promise-limit": "2.6.0",
     "react": "16.4.0",
     "react-dom": "16.4.0",
     "rimraf": "2.6.2",
