+++ taskfile.js
@@ -5,7 +5,7 @@ const mkdirp = require('mkdirp')
 const isWindows = /^win/.test(process.platform)
 
 export async function compile (task) {
-  await task.parallel(['bin', 'server', 'nextbuild', 'lib', 'client'])
+  await task.parallel(['bin', 'server', 'exportTask', 'nextbuild', 'lib', 'client'])
 }
 
 export async function bin (task, opts) {
@@ -21,6 +21,11 @@ export async function lib (task, opts) {
 export async function server (task, opts) {
   await task.source(opts.src || 'server/**/*.js').babel().target('dist/server')
   notify('Compiled server files')
+}
+
+export async function exportTask (task, opts) {
+  await task.source(opts.src || 'export/**/*.js').babel().target('dist/export')
+  notify('Compiled export files')
 }
 
 export async function nextbuild (task, opts) {
@@ -55,6 +60,7 @@ export default async function (task) {
   await task.watch('bin/*', 'bin')
   await task.watch('pages/**/*.js', 'copy')
   await task.watch('server/**/*.js', 'server')
+  await task.watch('export/**/*.js', 'exportTask')
   await task.watch('build/**/*.js', 'nextbuild')
   await task.watch('client/**/*.js', 'client')
   await task.watch('lib/**/*.js', 'lib')
